fix transfers
