+++ novawallet/Common/Services/RemoteSubscription/Substrate/BalanceRemoteSubscriptionService.swift
@@ -18,6 +18,21 @@ protocol BalanceRemoteSubscriptionServiceProtocol {
         queue: DispatchQueue?,
         closure: RemoteSubscriptionClosure?
     )
+
+    func attachToAssetBalance(
+        for accountId: AccountId,
+        chainAsset: ChainAsset,
+        queue: DispatchQueue?,
+        closure: RemoteSubscriptionClosure?
+    ) -> UUID?
+
+    func detachFromAssetBalance(
+        for subscriptionId: UUID,
+        accountId: AccountId,
+        chainAssetId: ChainAssetId,
+        queue: DispatchQueue?,
+        closure: RemoteSubscriptionClosure?
+    )
 }
 
 final class BalanceRemoteSubscriptionService: RemoteSubscriptionService {
@@ -52,6 +67,10 @@ final class BalanceRemoteSubscriptionService: RemoteSubscriptionService {
 
     private func createCacheKey(from accountId: AccountId, chainId: ChainModel.Id) -> String {
         "balances-\(accountId.toHex())-\(chainId)"
+    }
+
+    private func createCacheKey(from accountId: AccountId, chainAssetId: ChainAssetId) -> String {
+        "balances-\(accountId.toHex())-\(chainAssetId.chainId)-\(chainAssetId.assetId)"
     }
 
     private func prepareNativeAssetSubscriptionRequests(
@@ -329,6 +348,54 @@ extension BalanceRemoteSubscriptionService: BalanceRemoteSubscriptionServiceProt
         closure: RemoteSubscriptionClosure?
     ) {
         let cacheKey = createCacheKey(from: accountId, chainId: chainId)
+        detachFromSubscription(cacheKey, subscriptionId: subscriptionId, queue: queue, closure: closure)
+    }
+
+    func attachToAssetBalance(
+        for accountId: AccountId,
+        chainAsset: ChainAsset,
+        queue: DispatchQueue?,
+        closure: RemoteSubscriptionClosure?
+    ) -> UUID? {
+        let subscriptionSettingsList = prepareSubscriptionRequests(
+            from: accountId,
+            chainAsset: chainAsset,
+            transactionSubscription: nil
+        )
+
+        guard !subscriptionSettingsList.isEmpty else {
+            return nil
+        }
+
+        let cacheKey = createCacheKey(from: accountId, chainAssetId: chainAsset.chainAssetId)
+
+        let requests = subscriptionSettingsList.map(\.request)
+        let handlersStore = subscriptionSettingsList.reduce(
+            into: [String: RemoteSubscriptionHandlingFactoryProtocol]()
+        ) { accum, settings in
+            accum[settings.request.localKey] = settings.handlingFactory
+        }
+
+        let handlingFactory = BalanceRemoteSubscriptionHandlingProxy(store: handlersStore)
+
+        return attachToSubscription(
+            with: requests,
+            chainId: chainAsset.chain.chainId,
+            cacheKey: cacheKey,
+            queue: queue,
+            closure: closure,
+            subscriptionHandlingFactory: handlingFactory
+        )
+    }
+
+    func detachFromAssetBalance(
+        for subscriptionId: UUID,
+        accountId: AccountId,
+        chainAssetId: ChainAssetId,
+        queue: DispatchQueue?,
+        closure: RemoteSubscriptionClosure?
+    ) {
+        let cacheKey = createCacheKey(from: accountId, chainAssetId: chainAssetId)
         detachFromSubscription(cacheKey, subscriptionId: subscriptionId, queue: queue, closure: closure)
     }
 }
