+++ novawallet/Common/Services/ServiceCoordinator.swift
@@ -17,11 +17,11 @@ protocol ServiceCoordinatorProtocol: ApplicationServiceProtocol {
 
 final class ServiceCoordinator {
     let walletSettings: SelectedWalletSettings
-    let substrateBalancesService: AssetsUpdatingServiceProtocol
-    let evmAssetsService: AssetsUpdatingServiceProtocol
-    let evmNativeService: AssetsUpdatingServiceProtocol
+    let substrateBalancesService: AssetBalanceUpdatingServiceProtocol
+    let evmAssetsService: AssetBalanceUpdatingServiceProtocol
+    let evmNativeService: AssetBalanceUpdatingServiceProtocol
     let githubPhishingService: ApplicationServiceProtocol
-    let equilibriumService: AssetsUpdatingServiceProtocol
+    let equilibriumService: AssetBalanceUpdatingServiceProtocol
     let dappMediator: DAppInteractionMediating
     let proxySyncService: ProxySyncServiceProtocol
     let walletNotificationService: WalletNotificationServiceProtocol
@@ -30,11 +30,11 @@ final class ServiceCoordinator {
 
     init(
         walletSettings: SelectedWalletSettings,
-        substrateBalancesService: AssetsUpdatingServiceProtocol,
-        evmAssetsService: AssetsUpdatingServiceProtocol,
-        evmNativeService: AssetsUpdatingServiceProtocol,
+        substrateBalancesService: AssetBalanceUpdatingServiceProtocol,
+        evmAssetsService: AssetBalanceUpdatingServiceProtocol,
+        evmNativeService: AssetBalanceUpdatingServiceProtocol,
         githubPhishingService: ApplicationServiceProtocol,
-        equilibriumService: AssetsUpdatingServiceProtocol,
+        equilibriumService: AssetBalanceUpdatingServiceProtocol,
         proxySyncService: ProxySyncServiceProtocol,
         dappMediator: DAppInteractionMediating,
         walletNotificationService: WalletNotificationServiceProtocol,
@@ -122,7 +122,6 @@ extension ServiceCoordinator {
         let walletSettings = SelectedWalletSettings.shared
         let substrateStorageFacade = SubstrateDataStorageFacade.shared
 
-        let walletRemoteSubscription = WalletServiceFacade.sharedRemoteSubscriptionService
         let evmWalletRemoteSubscription = WalletServiceFacade.sharedEvmRemoteSubscriptionService
 
         let storageRequestFactory = StorageRequestFactory(
@@ -172,7 +171,7 @@ extension ServiceCoordinator {
         let equilibriumService = EquilibriumAssetBalanceUpdatingService(
             selectedAccount: walletSettings.value,
             chainRegistry: chainRegistry,
-            remoteSubscriptionService: walletRemoteSubscription,
+            remoteSubscriptionService: WalletServiceFacade.sharedEquillibriumRemoteSubscriptionService,
             repositoryFactory: SubstrateRepositoryFactory(storageFacade: substrateStorageFacade),
             storageRequestFactory: storageRequestFactory,
             eventCenter: EventCenter.shared,
