+++ novawallet/Modules/Transfer/TransferConfirm/TransferConfirmCrossChainViewFactory.swift
@@ -110,7 +110,7 @@ struct TransferConfirmCrossChainViewFactory {
 
         let repositoryFactory = SubstrateRepositoryFactory(storageFacade: storageFacade)
 
-        let walletRemoteSubscriptionService = WalletServiceFacade.sharedRemoteSubscriptionService
+        let walletRemoteSubscriptionService = WalletServiceFacade.sharedSubstrateRemoteSubscriptionService
 
         let walletRemoteSubscriptionWrapper = WalletRemoteSubscriptionWrapper(
             remoteSubscriptionService: walletRemoteSubscriptionService,
@@ -124,7 +124,9 @@ struct TransferConfirmCrossChainViewFactory {
         let senderResolutionFacade = ExtrinsicSenderResolutionFacade(userStorageFacade: UserDataStorageFacade.shared)
 
         let metadataHashOperationFactory = MetadataHashOperationFactory(
-            metadataRepositoryFactory: RuntimeMetadataRepositoryFactory(storageFacade: SubstrateDataStorageFacade.shared),
+            metadataRepositoryFactory: RuntimeMetadataRepositoryFactory(
+                storageFacade: SubstrateDataStorageFacade.shared
+            ),
             operationQueue: operationQueue
         )
