+++ novawallet/Modules/Transfer/TransferSetup/OnChain/TransferSetupPresenterFactory+OnChain.swift
@@ -118,7 +118,7 @@ extension TransferSetupPresenterFactory {
 
         let repositoryFactory = SubstrateRepositoryFactory(storageFacade: storageFacade)
 
-        let walletRemoteSubscriptionService = WalletServiceFacade.sharedRemoteSubscriptionService
+        let walletRemoteSubscriptionService = WalletServiceFacade.sharedSubstrateRemoteSubscriptionService
 
         let walletRemoteSubscriptionWrapper = WalletRemoteSubscriptionWrapper(
             remoteSubscriptionService: walletRemoteSubscriptionService,
